엔티티 맵핑, 테스트 추가
- 쿼리 오타 수정

+++ src/main/java/com/tathink/univa/repository/QuestionRepository.java
@@ -21,7 +21,7 @@ public class QuestionRepository {
 		return question;
 	}
 	
-	public Optional<Question> findById(Long id) {
+	public Optional<Question> findById(int id) {
 		Question question = em.find(Question.class, id);
 		return Optional.ofNullable(question);
 	}
